Move production information in publishconf
Previously we were setting SITEURL in the
pelicanconf.py . That was kind of problematic
because all the static resources link were prepended
by the site url, making it unknown when developping locally.
These kind of information are only valuable in production
not in development environment.

At deploying the publishconf will complete/override the
pelicanconf.